Create Shop Problem
There is a fair in a city that contains N shops. These shops are arranged in a straight line located at 1, 2, 3 ... N positions. M persons are visiting this ith person visit shops that are located between range Li and Ri inclusive.
The Shops that are visited by the most number of people are said to be more popular shop.
Determine the position of three most famous shops